+++ test/errors.test.js
@@ -12,26 +12,20 @@ describe('Errors', function () {
         const prevFileContent = fse.readFileSync('./lib/example/SampleFile.groovy').toString();
         const npmGroovyLintConfig = {
             source: prevFileContent,
-            insight: false,
             someUnknownParam: "lelamanul"
         };
-        let errRes;
-        try {
-            await new NpmGroovyLint(
-                npmGroovyLintConfig, {
-                jdeployRootPath: 'jdeploy-bundle'
-            }).run();
-        } catch (e) {
-            errRes = e;
-        }
-        assert(errRes != null, 'Error has been triggered');
-        assert(errRes.message.includes('Invalid option'), 'Invalid option detected');
+        const linter = await new NpmGroovyLint(
+            npmGroovyLintConfig, {
+            jdeployRootPath: 'jdeploy-bundle'
+        }).run();
+
+        assert(linter.status === 2, `Linter status is 2 (${linter.status} returned)`);
+        assert(linter.error.msg.includes('Invalid option'), 'Invalid option detected');
     });
 
     it('(API:source) should trigger a codenarc error', async () => {
         const npmGroovyLintConfig = {
             path: '/not/existing/path',
-            insight: false,
             output: 'txt',
             verbose: true
         };
@@ -40,14 +34,13 @@ describe('Errors', function () {
             jdeployRootPath: 'jdeploy-bundle'
         }).run();
 
-        assert(linter.status > 0, 'Linter status is > 0');
+        assert(linter.status === 2, `Linter status is 2 (${linter.status} returned)`);
         assert(linter.codeNarcStdErr && linter.codeNarcStdErr.includes('java.io.FileNotFoundException'), 'FileNotFoundException returned by CodeNarc');
     });
 
     it('(API:source) should trigger a codenarc error (--noserver)', async () => {
         const npmGroovyLintConfig = {
             path: '/not/existing/path',
-            insight: false,
             noserver: true
         };
         const linter = await new NpmGroovyLint(